Update README.md
To use this script:

Set the target_domain variable to the domain you want to emulate subdomains for (e.g., "vishnugit.com").
Create a text file called subdomains.txt and populate it with a list of subdomains you want to emulate, each on a new line. Following is an example list:
www
mail
blog
shop
app
api
cdn
dev
admin
test
stage
forum
secure
login
support
info
demo
Save the script to a file (e.g., subdomain_enu.sh) and make it executable using the command:
 chmod +x subdomain_enu.sh
Run the script using ./subdomain_enu.sh in the terminal.
The script will iterate over each subdomain in the wordlist file and perform a DNS lookup for it. If a DNS record is found for the subdomain, it will be displayed as a result. You can customize the script further to suit your specific needs, such as adding additional checks or modifying the output format.

Remember to replace "vishnugit.com" with the actual domain you want to emulate subdomains for, and create your own wordlist file containing the subdomains you want to test.